requirements.txt
## The following requirements were added by pip --freeze:
CoffeeScript==1.0.10
Django==1.7
PyExecJS==1.0.4
beautifulsoup4==4.3.2
coverage==3.7.1
dj-database-url==0.3.0
dj-redis-url==0.1.4
dj-static==0.0.6
django-rq==0.7.0
django-social-auth==0.7.28
django-tastypie==0.12.0
django-toolbelt==0.0.1
gnureadline==6.3.3
gunicorn==19.1.1
httplib2==0.9
ipython==2.2.0
lxml==3.3.5
mechanize==0.2.5
oauth2==1.5.211
psycopg2==2.5.3
python-dateutil==2.2
python-mimeparse==0.1.4
python-openid==2.2.5
readline==6.2.4.1
redis==2.10.3
rq==0.4.6
six==1.8.0
static3==0.5.1
wsgiref==0.1.2
